Refactoring the project
