+++ performance/page_performance.py
@@ -1,0 +1,111 @@
+# -*- coding: utf-8 -*-
+"""Tool for web-application performance analysis. It's based on Google's PageSpeed."""
+
+import requests
+
+
+class PagePerformance:
+    """
+    Performance analysis. It uses Google PageSpeed API under the hood.
+
+    Example of usage:
+      from python code:
+        test = PageSpeed("https://example.com/")
+      or from command line:
+        python3 performance.py https://example1.com http://example2.com http://example3.com
+
+    """
+    api_request_url = 'https://www.googleapis.com/pagespeedonline/v2/runPagespeed'
+    key = 'AIzaSyBXODcaGIeQDWqalKJTyzxOdqFbgAR8Vr8'
+
+    def __init__(self, url, strategy='all'):
+        self.url = url
+
+        if strategy == 'mobile':
+            self.res_mobile = requests.get(self.api_request_url,
+                                           params={'url': self.url,
+                                                   'strategy': 'mobile',
+                                                   'key': self.key})
+        elif strategy == 'desktop':
+            self.res_desktop = requests.get(self.api_request_url,
+                                            params={'url': self.url,
+                                                    'strategy': 'desktop',
+                                                    'key': self.key})
+        else:
+            self.res_mobile = requests.get(self.api_request_url,
+                                           params={'url': self.url,
+                                                   'strategy': 'mobile',
+                                                   'key': self.key})
+
+            self.res_desktop = requests.get(self.api_request_url,
+                                            params={'url': self.url,
+                                                    'strategy': 'desktop',
+                                                    'key': self.key})
+
+    def mobile_performance(self):
+        """
+        Mobile performance testing
+        :return: {'status_code': status code of the response ('int'),
+                  'id': url of the page under the performance test ('str'),
+                  'title': the title of the page under test ('str'),
+                  'speed': speed score of the page under test (score = 0...100, 'int'),
+                  'usability': usability (for mobile only) of the page under test (value = 0...100, 'int'),
+                   }
+
+        """
+        return {'status_code': self.res_mobile.json()["responseCode"],
+                'id': self.res_mobile.json()["id"],
+                'title': self.res_mobile.json()["title"],
+                'speed': self.res_mobile.json()["ruleGroups"]["SPEED"]["score"],
+                'usability': self.res_mobile.json()["ruleGroups"]["USABILITY"]["score"],  # for mobile only
+                }
+
+    def desktop_performance(self):
+        """
+        Desktop performance testing
+        :return: {'status_code': status code of the response ('int'),
+                  'id': url of the page under the performance test ('str'),
+                  'title': the title of the page under test ('str'),
+                  'speed': speed score of the page under test (score = 0...100, 'int'),
+                   }
+        """
+        return {'status_code': self.res_desktop.json()["responseCode"],
+                'id': self.res_desktop.json()["id"],
+                'title': self.res_desktop.json()["title"],
+                'speed': self.res_desktop.json()["ruleGroups"]["SPEED"]["score"],
+                }
+
+    def performance_adequacy(self, admissible_mobile=75, admissible_desktop=75):
+        """
+        The adequacy of mobile and desktop performance
+        :param admissible_mobile: the value of mobile performance that is admissible
+        :param admissible_desktop: the value of desktop performance that is admissible
+        :return: dictionary {"mobile": mobile_status,
+                             "desktop": desktop_status}
+        """
+        if self.res_mobile.status_code == 200:
+            value_mobile = self.res_mobile.json()["ruleGroups"]["SPEED"]["score"]
+            if value_mobile >= admissible_mobile:
+                mobile_status = (True, 'PASSED', value_mobile)
+            else:
+                mobile_status = (False, 'FAILED', value_mobile)
+        else:
+            value_mobile = 0
+            mobile_status = (False, 'ERROR', value_mobile)
+
+        if self.res_desktop.status_code == 200:
+            value_desktop = self.res_desktop.json()["ruleGroups"]["SPEED"]["score"]
+            if value_desktop >= admissible_desktop:
+                desktop_status = (True, 'PASSED', value_desktop)
+            else:
+                desktop_status = (False, 'FAILED', value_desktop)
+        else:
+            value_desktop = 0
+            desktop_status = (False, 'ERROR', value_desktop)
+
+        return {"mobile": mobile_status,
+                "desktop": desktop_status}
+
+
+__author__ = "Dmytro Bondarchuk"
+__version__ = "0.2.2"
